update v1 ignitionserver.go
