+++ control-plane-operator/controllers/hostedcontrolplane/ignitionserver/ignitionserver_test.go
@@ -178,48 +178,3 @@ func TestReconcileIgnitionServerServiceRoute(t *testing.T) {
 		})
 	}
 }
-
-func TestLookupMappedImage(t *testing.T) {
-
-	ocpOverrides := map[string][]string{
-		"quay.io/jparrill":    {"registry.hypershiftbm.lab:5000/jparrill"},
-		"quay.io/karmab":      {"registry.hypershiftbm.lab:5000/karmab"},
-		"quay.io/mavazque":    {"registry.hypershiftbm.lab:5000/mavazque"},
-		"quay.io/ocp-release": {"registry.hypershiftbm.lab:5000/openshift/release-images"},
-		"quay.io/openshift-release-dev/ocp-v4.0-art-dev":        {"registry.hypershiftbm.lab:5000/openshift/release"},
-		"registry.access.redhat.com/openshift4/ose-oauth-proxy": {"registry.hypershiftbm.lab:5000/openshift4/ose-oauth-proxy"},
-		"registry.redhat.io/lvms4":                              {"registry.hypershiftbm.lab:5000/lvms4"},
-		"registry.redhat.io/multicluster-engine":                {"registry.hypershiftbm.lab:5000/acm-d"},
-		"registry.redhat.io/openshift4":                         {"registry.hypershiftbm.lab:5000/openshift4"},
-		"registry.redhat.io/rhacm2":                             {"registry.hypershiftbm.lab:5000/acm-d"},
-		"registry.redhat.io/rhel8":                              {"registry.hypershiftbm.lab:5000/rhel8"},
-	}
-
-	tests := []struct {
-		name   string
-		in     string
-		expect string
-	}{
-		{
-			name:   "return the mirrorImage if it exists in the ocp overrides",
-			in:     "quay.io/openshift-release-dev/ocp-v4.0-art-dev@sha256:34acb17e45b80267919ea6a2f0e2b21e3f87db3c1e2443d2b3ac9afc0455ae07",
-			expect: "registry.hypershiftbm.lab:5000/openshift/release@sha256:34acb17e45b80267919ea6a2f0e2b21e3f87db3c1e2443d2b3ac9afc0455ae07",
-		},
-		{
-			name:   "return the same image if not in the overrides",
-			in:     "registry.ci.openshift.org/ocp/4.14-2023-09-05-120503@sha256:34acb17e45b80267919ea6a2f0e2b21e3f87db3c1e2443d2b3ac9afc0455ae07",
-			expect: "registry.ci.openshift.org/ocp/4.14-2023-09-05-120503@sha256:34acb17e45b80267919ea6a2f0e2b21e3f87db3c1e2443d2b3ac9afc0455ae07",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := lookupMappedImage(ocpOverrides, tt.in)
-			if err != nil {
-				t.Errorf("Unexpected error: %v", err)
-			}
-			if got != tt.expect {
-				t.Errorf("Wanted: %s, Got: %s", tt.expect, got)
-			}
-		})
-	}
-}
